move CompositionX to their respective X file
