+++ base/shared/src/main/scala/scalaz/tc/functor.scala
@@ -41,3 +41,10 @@ trait FunctorSyntax {
     def void(implicit ev: Functor[F]): F[Unit]            = ev.map[A, Unit](self)(_ => ())
   }
 }
+
+class CompositionFunctorClass[F[_], G[_]](implicit F: FunctorClass[F], G: FunctorClass[G])
+    extends FunctorClass[λ[α => F[G[α]]]] {
+
+  def map[A, B](fga: F[G[A]])(f: A => B): F[G[B]] =
+    F.map(fga)(G.map(_)(f))
+}
